fix vagrant provisioning to not reclone the project

+++ bootstrap.sh
@@ -28,15 +28,21 @@ then
   sudo -u postgres createuser -r -s -d vagrant
   sudo -u postgres createdb vagrant -O vagrant
 
-  # Clone the Meppit repo
-  cd /home/vagrant/
-  sudo -u vagrant git clone https://github.com/it3s/meppit.git meppit
-
   # Configure Meppit
+  cd /home/vagrant/
+  sudo -u vagrant ln -s /vagrant /home/vagrant/meppit
+
   cd meppit
-  sudo -u vagrant cp config/secrets.yml.sample config/secrets.yml
-  sudo -u vagrant cp config/database.yml.sample config/database.yml
-  sudo -u vagrant sed -i -e 's/username: postgres/username: vagrant/g' config/database.yml
+  if [ ! -f config/secrets.yml ]
+  then
+    sudo -u vagrant cp config/secrets.yml.sample config/secrets.yml
+  fi
+
+  if [ ! -f config/database.yml ]
+  then
+    sudo -u vagrant cp config/database.yml.sample config/database.yml
+    sudo -u vagrant sed -i -e 's/username: postgres/username: vagrant/g' config/database.yml
+  fi
 
   # Setup
   bundle
